Vim option 'updatetime' now 100ms.
This is for the live preview of latex files to update quicker. This
option was placed in the latex.vim ftplugin file so it is only
activated whilst editing latex files. NOTE! If other file types are being
edited in the same session the option remains at 100ms until the session
ends.